% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Plotting French name frquency since 1900}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_sex}{}
}
\description{
Plotting French name frquency since 1900
}
