% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{Read an excel file with multiple sheet into multiple lists}
\usage{
multi_excel(file_name)
}
\arguments{
\item{file_name}{full path including file name}
}
\value{
list of dataframes of excel sheets of a single xlsx file
}
\description{
Read an excel file with multiple sheet into multiple lists
}
\examples{
\dontrun{
 multi_excel("..")
}
}
