% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{Read csv files in a directory into a list}
\usage{
my_csv_reader(dir_path = getwd())
}
\arguments{
\item{dir_path}{directoy where the csv files are read}
}
\value{
a list of dataframes made up of csv
}
\description{
Read csv files in a directory into a list
}
\examples{
\dontrun{
my_csv_reader("..")
}
}
