% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Saving dataframe as csv in specified directory}
\usage{
save_as_csv(df, file_name)
}
\arguments{
\item{df}{dataframe}

\item{file_name}{name of the saved file}
}
\description{
Saving dataframe as csv in specified directory
}
\examples{
save_as_csv(iris, "iris_data")
}
